Convert audio files to target settings on load

+++ include/audiosource.h
@@ -132,7 +132,7 @@ protected:
 private:
     ValueObject* resource;
 
-    double index;
+    unsigned int index;
 
 };
